chore(TenantSubscriptionModes & UserSubscriptionModes) Enhanced component structure and typing
