+++ components/HomePageInterface/components/SubscriptionCardsFooterDisclaimer/index.ts
@@ -1,0 +1,1 @@
+export { default as SubscriptionCardsFooterDisclaimer } from "./SubscriptionCardsFooterDisclaimer";
