+++ components/HomePageInterface/components/SubscriptionModesSectionBackground/index.ts
@@ -1,0 +1,1 @@
+export { default as SubscriptionModesSectionBackground } from "./SubscriptionModesSectionBackground";
