+++ components/HomePageInterface/components/SubscriptionModesSectionFooterDisclaimer/index.ts
@@ -1,0 +1,1 @@
+export { default as SubscriptionModesSectionFooterDisclaimer } from "./SubscriptionModesSectionFooterDisclaimer";
