+++ components/HomePageInterface/components/SubscriptionModesSectionTitle/index.ts
@@ -1,0 +1,1 @@
+export { default as SubscriptionModesSectionTitle } from "./SubscriptionModesSectionTitle";
