+++ components/HomePageInterface/components/TenantSubscriptionModes/TenantSubscriptionModes.tsx
@@ -1,75 +1,29 @@
 "use client";
 import { Tenant } from "@/context/TenantContext/types";
 import TenantSubscriptionCard from "./components/TenantSubscriptionCard/TenantSubscriptionCard";
-
-const TENANT_PLANS = {
-  tenant: {
-    tenantScopedQuota: 10,
-    productLimit: 100,
-    monthlySpinsLimit: 200000,
-  },
-  tenantPro: {
-    tenantScopedQuota: 20,
-    productLimit: 250,
-    monthlySpinsLimit: 500000,
-  },
-  tenantPremium: {
-    tenantScopedQuota: 30,
-    productLimit: 500,
-    monthlySpinsLimit: 850000,
-  },
-};
+import { TENANT_PLANS, tenantPlans } from "../../utils/tenantHelpers";
+import { SubscriptionModesSectionFooterDisclaimer } from "../SubscriptionModesSectionFooterDisclaimer";
+import { SubscriptionModesSectionTitle } from "../SubscriptionModesSectionTitle";
+import { SubscriptionModesSectionBackground } from "../SubscriptionModesSectionBackground";
 
 const TenantSubscriptionModes = ({ tenant }: { tenant?: Tenant | null }) => {
-  const CURRENT_TENANT_PLAN = tenant?.subscriptionMode; // 🔥 mock for now
-
-  const plans = [
-    {
-      id: "tenant",
-      name: "TENANT",
-      price: "R$299,",
-    },
-    {
-      id: "tenantPro",
-      name: "TENANT PRO",
-      price: "R$499,",
-    },
-    {
-      id: "tenantPremium",
-      name: "TENANT PREMIUM",
-      price: "R$799,",
-      highlight: true,
-    },
-  ];
-
+  const CURRENT_TENANT_PLAN = tenant?.subscriptionMode;
   const tenantMaxPlan = tenant?.subscriptionMode === "tenantPremium";
+  const title = "PLANOS PARA PARCEIROS";
+  const subTitle =
+    "Escale sua presença na Promobet, aumente seus giros e amplie sua vitrine de produtos.";
+  const message =
+    "Assinaturas para parceiros serão processadas via Stripe. É possível alterar ou cancelar seu plano a qualquer momento.";
 
   if (tenantMaxPlan) {
     return null;
   }
 
   return (
-    <section
-      id="subscription-modes-tenant"
-      className="relative md:mx-2 my-20  overflow-hidden bg-gradient-to-br from-[#0b0f1f] via-[#141b3a] to-[#1d1147] text-white p-16"
-    >
-      {/* Glow */}
-      <div className="absolute inset-0 flex items-center justify-center pointer-events-none">
-        <div className="w-[700px] h-[700px] bg-purple-500 opacity-10 blur-[180px] rounded-full" />
-      </div>
-
-      <div className="relative z-10 text-center mb-16 cursor-default">
-        <h2 className="text-5xl font-extrabold tracking-widest bg-gradient-to-r text-[#84e9e4] ">
-          PLANOS PARA PARCEIROS
-        </h2>
-        <p className="mt-6 text-gray-300 max-w-2xl mx-auto">
-          Escale sua presença na Promobet, aumente seus giros e amplie sua
-          vitrine de produtos.
-        </p>
-      </div>
-
+    <SubscriptionModesSectionBackground>
+      <SubscriptionModesSectionTitle title={title} subTitle={subTitle} />
       <div className="grid gap-8 md:grid-cols-3 max-w-7xl mx-auto">
-        {plans.map((plan) => {
+        {tenantPlans.map((plan) => {
           const config = TENANT_PLANS[plan.id as keyof typeof TENANT_PLANS];
 
           return (
@@ -82,12 +36,8 @@ const TenantSubscriptionModes = ({ tenant }: { tenant?: Tenant | null }) => {
           );
         })}
       </div>
-
-      <div className="mt-16 text-center text-xs text-gray-500 max-w-3xl mx-auto cursor-default">
-        Assinaturas para parceiros serão processadas via Stripe. É possível
-        alterar ou cancelar seu plano a qualquer momento.
-      </div>
-    </section>
+      <SubscriptionModesSectionFooterDisclaimer message={message} />
+    </SubscriptionModesSectionBackground>
   );
 };
