+++ components/HomePageInterface/components/TenantSubscriptionModes/components/TenantSubscriptionCardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as TenantSubscriptionCardHeader } from "./TenantSubscriptionCardHeader";
