+++ components/HomePageInterface/components/TenantSubscriptionModes/components/TenantSubscriptionCardPerks/index.ts
@@ -1,0 +1,1 @@
+export { default as TenantSubscriptionCardPerks } from "./TenantSubscriptionCardPerks";
