+++ components/HomePageInterface/components/UserSubscriptionModes/components/UserSubscriptionCard/UserSubscriptionCard.tsx
@@ -1,32 +1,17 @@
-import UserSubscriptionButton from "../UserSubscriptionButton/UserSubscriptionButton";
+import { UserSubscriptionCardProps } from "@/components/HomePageInterface/types";
 import { useRouter } from "next/navigation";
+import { SubscriptionCardsFooterDisclaimer } from "../../../SubscriptionCardsFooterDisclaimer";
+import { UserSubscriptionButton } from "../UserSubscriptionButton";
+import { UserSubscriptionCardPerks } from "../UserSubscriptionCardPerks";
+import { UserSubscriptionCardHeader } from "../UserSubscriptionCardHeader";
 
 const UserSubscriptionCard = ({
   plan,
   config,
   userPlan,
-}: {
-  plan:
-    | {
-        id: string;
-        name: string;
-        price: string;
-        highlight?: undefined;
-      }
-    | {
-        id: string;
-        name: string;
-        price: string;
-        highlight: boolean;
-      };
-  config: {
-    global: number;
-    tenantMultiplier: number;
-    monthlyGlobalTenantsQuota: number;
-    weeklyGlobalTenantsQuota: number;
-  };
-  userPlan: string | undefined;
-}) => {
+}: UserSubscriptionCardProps) => {
+  const message =
+    "Os limites podem ser extendidos através de pacotes vendidos separadamente.";
   const router = useRouter();
   const handleSubscribe = (planId: string) => {
     console.log("Subscribe to:", planId);
@@ -52,120 +37,14 @@ const UserSubscriptionCard = ({
         </div>
       )}
 
-      <h3 className="text-2xl font-bold tracking-wide mb-4 text-amber-500 cursor-default">
-        {plan.name}
-      </h3>
-
-      <div className="flex space-x-1 items-center cursor-default  mb-6 bg-gradient-to-r from-[#84e9e4] to-amber-500 hover:opacity-90 bg-clip-text text-transparent">
-        <div className="text-4xl font-extrabold">{plan.price}</div>
-        <div className="text-xl font-bold mt-4">99</div>
-      </div>
-
-      <ul className="space-y-2 text-gray-300 mb-8 cursor-default">
-        <li>
-          <div className="flex space-x-1 items-center justify-between border-b align-text-top">
-            <div className="pt-2">
-              <span className="font-semibold tracking-widest text-slate-300 text-md drop-shadow-2xl">
-                Giros por Dia
-              </span>
-            </div>
-            <div className="flex space-x-1 items-center">
-              <div className="flex space-x-1 items-center ">
-                <span className="text-3xl font-bold tracking-widest text-[#84e9e4]">
-                  {config.global}
-                </span>
-              </div>
-              <div className="pt-2">
-                <span className="text-xs font-semibold tracking-widest text-[#84e9e4]">
-                  x
-                </span>
-              </div>
-            </div>
-          </div>
-        </li>
-        <div>
-          <span className="text-xs font-semibold tracking-widest text-slate-500">
-            Parceiros
-          </span>
-          <div className="bg-slate-50/2 px-2 pb-1">
-            <li>
-              <div className="flex space-x-1 items-center justify-between">
-                <div className="pt-2">
-                  <span className="font-semibold tracking-widest text-slate-300 text-md drop-shadow-2xl">
-                    Giros Extras
-                  </span>
-                </div>
-                <div className="flex space-x-1 items-center">
-                  <div className="flex space-x-1 items-center ">
-                    <span className="text-2xl font-bold tracking-widest text-[#84e9e4] drop-shadow-2xl">
-                      {config.tenantMultiplier > 0
-                        ? `${config.tenantMultiplier === 0.5 ? " 10 " : " 30 "}`
-                        : "Não incluso"}
-                      {""}
-                    </span>
-                  </div>
-                  <div className="pt-2">
-                    <span className="text-xs font-semibold tracking-widest text-[#84e9e4]">
-                      x
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </li>
-            <li>
-              <div className="flex space-x-1 items-center justify-between">
-                <div className="pt-2">
-                  <span className="font-semibold tracking-widest text-slate-300 text-md drop-shadow-2xl">
-                    Limite de Giros
-                  </span>
-                </div>
-                <div className="flex space-x-1 items-center">
-                  <div className="flex space-x-1 items-center ">
-                    <span className="text-2xl font-bold tracking-widest text-[#84e9e4] drop-shadow-2xl">
-                      {config.monthlyGlobalTenantsQuota}
-                    </span>
-                  </div>
-                  <div className="pt-2">
-                    <span className="text-xs font-semibold tracking-widest text-[#84e9e4]">
-                      x
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </li>
-            <li>
-              <div className="flex space-x-1 items-center justify-between">
-                <div className="pt-2">
-                  <span className="font-semibold tracking-widest text-slate-300 text-xs drop-shadow-2xl">
-                    Limite por semana
-                  </span>
-                </div>
-                <div className="flex space-x-1 items-center">
-                  <div className="flex space-x-1 items-center pt-2">
-                    <span className="text-sm font-bold tracking-widest text-[#84e9e4] drop-shadow-2xl">
-                      {config.weeklyGlobalTenantsQuota}
-                    </span>
-                  </div>
-                  <div className="pt-1">
-                    <span className="text-xs font-semibold tracking-widest text-[#84e9e4]">
-                      x
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </li>
-          </div>
-        </div>
-      </ul>
+      <UserSubscriptionCardHeader plan={plan} />
+      <UserSubscriptionCardPerks config={config} />
       <UserSubscriptionButton
         handleSubscribe={handleSubscribe}
         isUserPlan={isUserPlan}
         plan={plan}
       />
-      <div className="italic pt-4 text-center text-xs text-gray-500 max-w-3xl mx-auto cursor-default">
-        Os limites podem ser extendidos através de pacotes vendidos
-        separadamente.
-      </div>
+      <SubscriptionCardsFooterDisclaimer message={message} />
     </div>
   );
 };
