+++ components/HomePageInterface/components/UserSubscriptionModes/components/UserSubscriptionCardPerks/index.ts
@@ -1,0 +1,1 @@
+export { default as UserSubscriptionCardPerks } from "./UserSubscriptionCardPerks";
